rbperf: fix ruby_version symbol lookup
ruby_version looks up for the ruby_version symbol
and reads it from the process binary. before, it
read an offset with a fixed length (5) that resulted
in returning 2.4.1 instead of 2.4.10.

this patch fixes bcc.bcc_elf_foreach.callback signature
in symbol_address and uses symbol payload length
to do the correct reading.

then ruby_version now returns encoded and stripped
string instead of a bytes sequence for the version.
it also now has a different signature returning a
tuple of an (addr, length).